Added the functionality in for the show notifications and keep playlist settings

+++ app/src/main/java/com/jb/dailyplay/managers/DailyPlayMusicManager.java
@@ -132,6 +132,10 @@ public class DailyPlayMusicManager {
             return;
         }
 
+        if (DailyPlaySharedPrefUtils.getKeepPlayList()) {
+            return;
+        }
+
         for(SongFile downloadedFile : mDownloadedFiles) {
             File file = downloadedFile.getFile();
             file.delete();
